requirements.txt
matplotlib==3.9.0
numpy==1.24.3
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
Pillow==10.3.0
seaborn==0.13.2
torch==2.2.0+cu118
torcheval==0.0.7
torchvision==0.17.0+cu118
tqdm==4.66.1
